requirements.txt
annotated-types==0.7.0
asgiref==3.7.2
async-timeout==4.0.3
cachetools==5.5.2
certifi==2023.11.17
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==4.2.7
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==20.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
inflection==0.5.1
packaging==23.2
pillow==10.2.0
proto-plus==1.26.1
protobuf==5.29.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.8.0
pyparsing==3.2.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
stripe==7.12.0
toposort==1.10
tqdm==4.67.1
#typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
watchdog==4.0.0
